+++ src/card/sets/y2021/modern_horizons_2.rs
@@ -703,6 +703,7 @@ static RAGAVAN_CONNECTS: [EffectDef; 2] = [
         duration: ExilePlayDurationDef::ThisTurn,
         spend_any_color: false,
         play_condition: None,
+        cast_only: true,
     },
 ];