+++ src/card/sets/y2024/magic_foundations.rs
@@ -93,6 +93,7 @@ static KELLAN_INVESTIGATES: AbilityDef = AbilityDef::triggered(
         duration: ExilePlayDurationDef::ThisTurn,
         spend_any_color: false,
         play_condition: None,
+        cast_only: false,
     },
 );